refactor: MoviePosterImageLoader 클래스의 메서드를 MovieInfoDataLoader로 통합 후 삭제
